apps/ml/pyproject.toml
[project]
name = "apidelta-ml"
version = "0.1.0"
description = "APIDelta ML classifier for OpenAPI diffs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "tenacity>=8.2",
    "pyyaml>=6.0",
    "orjson>=3.10",
    "click>=8.1",
    "rich>=13.7",
]

[project.optional-dependencies]
ml = [
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "sentence-transformers>=2.7",
    "mapie>=0.8",
]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.3",
    "respx>=0.20",
]

[project.scripts]
apidelta-ml = "apidelta_ml.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apidelta_ml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers"
markers = [
    "network: test hits the network (slow, skipped with `make test-unit`)",
    "slow: slow test (>5s)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
